+++ src/main/java/mods/railcraft/common/blocks/tracks/behaivor/SpeedController.java
@@ -22,6 +22,7 @@ import net.minecraft.entity.item.EntityMinecart;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.IBlockAccess;
 import net.minecraft.world.World;
+
 import org.jetbrains.annotations.Nullable;
 
 /**
@@ -78,7 +79,7 @@ public enum SpeedController {
             BlockRailBase.EnumRailDirection dir = TrackTools.getTrackDirection(world, pos, cart);
             if (dir.isAscending())
                 return HighSpeedTools.SPEED_SLOPE;
-            return (float) HighSpeedTools.speedForNextTrack(world, pos, 0, cart);
+            return HighSpeedTools.speedForNextTrack(world, pos, 0, cart);
         }
     },
     REINFORCED {
